requirements.txt
# Core dependencies
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0

# XML parsing
lxml>=4.9.0
xmltodict>=0.13.0

# Template engine
jinja2>=3.1.0

# CLI and configuration
click>=8.1.0
python-decouple>=3.8
pyyaml>=6.0

# Logging and monitoring
loguru>=0.7.0

# Code formatting and validation
black>=23.0.0
isort>=5.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data validation
pydantic>=2.0.0

# Graph analysis
networkx>=3.1 